Add support for styling all gutters on a given line
BREAKING: `GutterMarker.at` is now deprecated (use \`GutterMarker.range\`).

FEATURE: The new `gutterLineClass` facet can now be used to style all
gutter elements for a given line.

FEATURE: The package now exports a `highlightActiveLineGutter`
extension that can be used to add a style to the gutter for the active
line.